network: fixed incorrect timeout parameter for NET_FlushPacketQueue() win32: added netmask detection function for ipv4 addresses misc: fixed various compiler warnings

+++ code/qcommon/msg.c
@@ -219,7 +219,7 @@ static int MSG_ReadBits( msg_t *msg, int bits ) {
 
 	if ( sgn && bits < 32 ) {
 		if ( value & ( 1 << ( bits - 1 ) ) ) {
-			value |= -1 ^ ( ( 1 << bits ) - 1 );
+			value |= -1 ^ ( ( 1U << bits ) - 1 );
 		}
 	}
 
@@ -290,7 +290,7 @@ void MSG_WriteString( msg_t *sb, const char *s ) {
 
 	for ( i = 0 ; i < l; i++ ) {
 		// get rid of 0x80+ and '%' chars, because old clients don't like them
-		if ( s[i] & 0x80 || s[i] == '%' )
+		if ( (byte)s[i] & 0x80 || s[i] == '%' )
 			v = '.';
 		else
 			v = s[i];
@@ -312,7 +312,7 @@ void MSG_WriteBigString( msg_t *sb, const char *s ) {
 
 	for ( i = 0 ; i < l ; i++ ) {
 		// get rid of 0x80+ and '%' chars, because old clients don't like them
-		if ( s[i] & 0x80 || s[i] == '%' )
+		if ( (byte)s[i] & 0x80 || s[i] == '%' )
 			v = '.';
 		else
 			v = s[i];
